Core Library Implementation & Library Patch 2
Introductions:
🟢 Added Print Fuction to AMPLE::Core.

Changes:
🔵 Fixed Base Libraries.
🔵 Fixed Memory and Register errors.
🔵 Fixed Mac OS Library loading problem.

Known issues:
Known issues:
🔴 If a swap partition is used, memset should not be used as it requires the memory to be set to 0x00. As a result run-time speeds will suffer.
🔴 Has not been tested in MacOS.

Future improvements:
🟡 Library installation.
🟡 Will introduce Programs integrity in Load Program.
🟡 Will implement Core Library.
🟡 Extend I/O Library.
🟡 Fix AMPLE.h warnings so they are not treated as errors.
🟡 Introduce powershell script for compiling an testing AMPLE (Introduced shell for Linux and Mac OS).

+++ Libraries/Memory/include/Memory.h
@@ -8,10 +8,10 @@ class Memory
 {
 protected:
     void* _memory;
-    size_t _size;
+    uint64_t _size;
 public:
-    EXPORT explicit Memory(size_t size);
-    EXPORT virtual void ResizeMemory(size_t size);
+    EXPORT explicit Memory(uint64_t size);
+    EXPORT virtual void ResizeMemory(uint64_t size);
     EXPORT virtual void* Get(uint64_t pos);
     EXPORT virtual ~Memory();
 };
